DRY plugin code, into a common abstract view
